requirements.txt
tuya-connector-python~=0.1.2
